Support for A500+.
These changes let A314 work in an A500+, providing an additional 1MB of chip ram for a total of 2MB chip.
The SPI protocol had to be modified so that addresses are sent as 20 bits instead of 19 bits, so that the entire 1MB of SRAM can be addressed.
The GetA314MemBase() API function exposed by a314.device was replaced with a more general TranslateAddressA314() function. All applications that previously used GetA314MemBase() had to be updated. This fixes #12.

At this time a small change needs to be made in HDL/a314_top.v at line 107, and the Verilog then needs to be recompiled and written to the FPGA, in order to use A314 in an A500+. The reason is that when A314 is used in a regular A500 the A314 should not respond to accesses when the RAS0 pin is asserted, but there is currently no way in hardware to disconnect RAS0 (#11). Later hardware revisions may address this.