Do dynamic bank mapping detection
This change allows the Amiga to detect how Amiga addresses are mapped to
A314 addresses, instead of guessing based on what memory blocks are present.

If the A314 has an older firmware, as seen by reporting firmware version 0,
the old heuristical method will be used. Updating the firmware is encouraged.