Added remote-mouse service
The remote-mouse service takes input events from a mouse connected to the Pi
from /dev/input/mice, and feeds the events to input.device on the Amiga.

You may want to disable the mouse in X11 if your Pi is running a GUI.
Use the command "xinput list" to find the id of your mouse to disable,
and then use "xinput --disable <id>".