Added support for A314-600 to master branch
Previously there was a separate branch for the A600 version of A314, but this
functionality has now been merged into the master branch.

A new revision, a314_600, has been added to the Quartus project.
A macro, is_a600, is used to control the compilation in a few places, mostly in
the top entity a314_top.v.

A new SPI protocol command has been added, SPI_PROTOCOL_VERSION. On older HDL
this command returns the default value zero, and this version returns one.
The a314d uses this protocol version number to be able to communicate with both
the current and older versions of the HDL.

The firmware version number has been replaced with a firmware flags value,
that consists of multiple bit flags that describes what capabilities the
firmware has.

Ideally this commit should probably have been multiple smaller commits, but to
save time I didn't bother to split it up.