Add support for comments and file/directory metadata
Amiga metadata that can not be mapped to any Linux equivalent can be
stored in a metadata file. Using a colon (:), an illegal filename
character on the Amiga, in the metadata filename assures that there
can be no collisions with any Amiga files.

The metadata file will only be created if the metadata changes from
the default.