Change default target frequency to 90 MHz
This is the frequency of the clock that the Raspberry Pi outputs,
that acts as the main clock driving the logic in the CPLD.

Note that a target frequency of 90 MHz is rounded down to 85.7 MHz.
And this assumes that the PLLC clock in the Pi is running at 1200 MHz.

If the Pi is using dynamic overclocking, which is the default unless
the line force_turbo=1 is in /boot/config.txt, then the Pi will
lower the PLLC clock to 1000 MHz when the ARM cores are inactive.
This lowers the clock going to the CPLD down to 71.4 MHz.

From the testing I've done it seems that the logic works fine also
with this lower frequency.