Add PiHid service
The PiHid service is an extension of the RemoteMouse service,
as the PiHid service allows both mice and keyboards to be connected
to the Raspberry Pi, and the generated input events are sent to
input.device on the Amiga.

The service supports hot-plugging so that a mouse/keyboard can be
added or removed at any time.