Update makefile for CP and TD models, add a314net.service
Using MODEL=TD or MODEL=CP make can now build and install the A314
Linux binaries and support files for both the trapdoor and clockport
hardware.

This change also adds a systemd a314net service, replacing the
recommended commands to add to rc.local for sharing the Internet
connection with the Amiga.