Add dist make target to build installation packages
Using the dist target, a self-extracting and installing package can be
built. This package can then be distributed to end user for a very
easy installation. The installation script installs the A314 binaries
and any support files. It can also optionally install dependencies and
perform necessary changes to configuration files in /boot.